requirements.txt
requests
tweepy
python-dotenv
firebase-admin
python-socketio[client]
tensorflow==2.2
ktrain